+++ src/pipa.cpp
@@ -1,6 +1,6 @@
 #include "Pipa.hpp"
 
-Pipa::Pipa(float pos_x, float pos_y) : Coletavel(pos_x, pos_y, "pipa") {};
+Pipa::Pipa(double pos_x, double pos_y) : Coletavel(pos_x, pos_y, "pipa") {};
 
 Pipa::~Pipa() {};
